restructuring and install tags
