+++ cob_torso_calibration/CMakeLists.txt
@@ -3,12 +3,18 @@ project(cob_torso_calibration)
 
 find_package(catkin REQUIRED COMPONENTS)
 
-#catkin_python_setup()
+catkin_python_setup()
 
 catkin_package()
 
 
 ###install tags
-#install(DIRECTORY launch
-  #DESTINATIN ${CATKIN_PACKAGE_SHARE_DESTINATION}
-#)
+install(DIRECTORY launch
+  DESTINATIN ${CATKIN_PACKAGE_SHARE_DESTINATION}
+)
+
+install(PROGRAMS src/calibration_lower_upper_neck.py
+                 src/update_cob_torso_calibration_urdf.py
+  DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}/src
+)
+
