+++ cob_torso_calibration/setup.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+   ##  don't do this unless you want a globally visible script
+   packages=['cob_torso_calibration'],
+   package_dir={'': 'src'}
+)
+
+setup(**d)
