+++ cob_torso_calibration/src/cob_torso_calibration/__init__.py
@@ -1,0 +1,1 @@
+from torso_state_calculation import *
